refactor: create a jira module & add env variables for jira service
