+++ .env.example
@@ -1,6 +1,9 @@
 NODE_ENV = development
 CPV_BASEHOST = http://localhost
 CPV_BASEPATH = /cpv
+CPV_JIRA_BASE_URL = https://domain.atlassian.net
+CPV_JIRA_API_USERNAME = 
+CPV_JIRA_API_TOKEN = 
 CPV_CONFLUENCE_BASE_URL = https://domain.atlassian.net
 CPV_CONFLUENCE_API_USERNAME = 
 CPV_CONFLUENCE_API_TOKEN =
