+++ src/http/http.module.ts
@@ -8,7 +8,6 @@ import {
 } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import Config from '../config/config';
-import { JiraService } from './jira.service';
 
 @Module({
   imports: [
@@ -26,8 +25,7 @@ import { JiraService } from './jira.service';
       inject: [ConfigService],
     }),
   ],
-  providers: [JiraService],
-  exports: [BaseHttpModule, JiraService],
+  exports: [BaseHttpModule],
 })
 export class HttpModule implements OnModuleInit {
   constructor(private readonly httpService: HttpService) {}
