+++ src/proxy-page/proxy-page.step.ts
@@ -1,4 +1,4 @@
-import { JiraService } from 'src/http/jira.service';
+import { JiraService } from 'src/jira/jira.service';
 import { ContextService } from '../context/context.service';
 
 export interface Step {
